requirements.txt
certifi==2020.6.20
click==7.1.2
gitdb==4.0.5
GitPython==3.1.3
numpy==1.19.0
pandas==1.0.5
pygame==1.9.6
python-box==4.2.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
smmap==3.0.4
toml==0.10.1
wincertstore==0.2
